+++ qcfractal/interface/tests/test_visualization.py
@@ -8,6 +8,7 @@ import requests
 from . import portal
 
 try:
+    import plotly
     _has_ploty = True
 except ModuleNotFoundError:
     _has_ploty = False
