+++ setup.py
@@ -52,7 +52,8 @@ if __name__ == "__main__":
                 "qcfractal-server=qcfractal.cli.qcfractal_server:main",
                 "qcfractal-manager=qcfractal.cli.qcfractal_manager:main",
                 "qcfractal-dashboard=qcfractal.cli.qcfractal_dashboard:main",
-            ]
+            ],
+            "pytest11": ["qcfractal_testing=qcfractal.testing"]
         },
         extras_require={
             'docs': [
